Destructure the HostGameScreen component
